refactor: 카드와 관련된 도메인 객체들을 한 패키지에 정리하기
