+++ src/test/kotlin/blackjack/playingcard/SymbolTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.playingcard
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
